+++ com.oracle.truffle.r.ffi.impl/src/com/oracle/truffle/r/ffi/impl/javaGD/BufferedImageGD.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2020, Oracle and/or its affiliates. All rights reserved.
+ * Copyright (c) 2020, 2022, Oracle and/or its affiliates. All rights reserved.
  * DO NOT ALTER OR REMOVE COPYRIGHT NOTICES OR THIS FILE HEADER.
  *
  * This code is free software; you can redistribute it and/or modify it
@@ -43,9 +43,10 @@ public final class BufferedImageGD extends GDInterface {
     }
 
     @Override
-    public void gdOpen(double width, double height) {
+    public boolean gdOpen(double width, double height) {
         super.gdOpen(width, height);
         c = new BufferedImageContainer((int) width, (int) height, fileType, fileNameTemplate, quality);
+        return true;
     }
 
 }