+++ com.oracle.truffle.r.runtime/src/com/oracle/truffle/r/runtime/Collections.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2013, 2021, Oracle and/or its affiliates. All rights reserved.
+ * Copyright (c) 2013, 2022, Oracle and/or its affiliates. All rights reserved.
  * DO NOT ALTER OR REMOVE COPYRIGHT NOTICES OR THIS FILE HEADER.
  *
  * This code is free software; you can redistribute it and/or modify it
@@ -558,19 +558,23 @@ public final class Collections {
             return result;
         }
 
-        /**
-         * Removes the last element.
-         */
         @ExportMessage
         public Object pop() {
             if (size <= 0) {
-                CompilerDirectives.transferToInterpreter();
-                throw new ArrayIndexOutOfBoundsException("cannot pop from empty list");
+                throw CompilerDirectives.shouldNotReachHere("Cannot pop from empty list");
             }
             size--;
             Object removed = data[size];
             data[size] = null;
             return removed;
+        }
+
+        @ExportMessage
+        public Object peek() {
+            if (size <= 0) {
+                throw CompilerDirectives.shouldNotReachHere("Cannot peek empty list");
+            }
+            return data[size - 1];
         }
 
         public int size() {
@@ -618,7 +622,15 @@ public final class Collections {
 
         public abstract void push(Object stack, Object value);
 
+        /**
+         * Removes the last element.
+         */
         public abstract Object pop(Object stack);
+
+        /**
+         * Shows the last element without removing it.
+         */
+        public abstract Object peek(Object stack);
     }
 
 }