+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # 22.1.0
 * Improved performance of the `order` and `rank` builtin functions
+* Use JavaGD as the default graphical subsystem.
+  * Deprecate `--R.UseInternalGridGraphics` option.
+  * The FastR's graphical subsystem is now mostly compatible with GNU-R's, i.e., most functions from `graphics`, `grid`, and `grDevices` base packages are now supported.
+  * Display lists are fully implemented.
+  * Supported devices: SVG, PNG, JPEG, BMP, AWT.
+  * See [graphics docs](./documentation/dev/graphics.md).
 
 # 22.0.0
 * Adopted [NodeLibrary](https://www.graalvm.org/truffle/javadoc/com/oracle/truffle/api/interop/NodeLibrary.html).